pyproject.toml
[project]
name = "cdvae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.7.0",
    "torch_geometric",
    "pyg_lib",
    "torch_scatter",
    "torch_sparse",
    "torch_cluster",
    "torch_spline_conv",
    "pymatgen",
    "pytorch-lightning>=2.5.2",
    "wandb>=0.21.0",
    "hydra-core>=1.3.2",
    "scikit-learn>=1.7.1",
    "p-tqdm>=1.4.2",
    "dotenv>=0.9.9",
    "smact>=3.2.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128"},
]
pyg_lib = [
    { index = "geometry-cu128"},
]
torch_scatter = [
    { index = "geometry-cu128"},
]
torch_sparse = [
    { index = "geometry-cu128"},
]
torch_cluster = [
    { index = "geometry-cu128"},
]
torch_spline_conv = [
    { index = "geometry-cu128"},
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "geometry-cu128"
url = "https://data.pyg.org/whl/torch-2.7.0+cu128.html"
explicit = true
format = "flat"

[[tool.uv.index]]
name = "geometry-cpu"
url = "https://data.pyg.org/whl/torch-2.7.0+cpu.html"
explicit = true
format = "flat"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["cdvae", "cdvae.*"]  # 只包含 cdvae 包
exclude = ["data", "conf", "assets", "tests*"]  # 排除其他目录
